+++ Packages/com.shijbey.tdrs/Runtime/SocialAgent.cs
@@ -262,7 +262,7 @@ namespace TDRS
 		{
 			string statName = nameAndValue.Item1;
 			float value = nameAndValue.Item2;
-			Engine.DB.Insert($"{UID}.stat.{statName}!{value}");
+			Engine.DB.Insert($"{UID}.stats.{statName}!{value}");
 			if (OnStatChange != null) OnStatChange.Invoke(statName, value);
 		}