+++ CHANGELOG.md
@@ -4,8 +4,25 @@ All notable changes to this project will be documented in this file.
 
 The format is based on <https://common-changelog.org/>, and this project adheres mostly to Semantic Versioning. However, all releases before 1.0.0 have breaking changes between minor-version updates.
 
+## [1.1.0] - 2024-01-25
+
+### Changed
+
+- **breaking:** Rename `SocialEventType` to `SocialEvent`
+- **breaking:** Pluralized `stats` within the RePraxis database
+
+### Added
+
+- Users can create traits and social events using ScriptableObjects
+- `DontDestroyOnLoad` option for `SocialEngine`
+
+### Fixed
+
+- Corrected `EffectName` property of `RemoveAgentTraitFactory` to `RemoveAgentTrait`
+
 ## [1.0.0] - 2024-01-11
 
 _Initial release._
 
 [1.0.0]: https://github.com/ShiJbey/Unity-TDRS/releases/tag/v1.0.0
+[1.1.0]: https://github.com/ShiJbey/Unity-TDRS/releases/tag/v1.1.0